+++ zcash_client_backend/src/data_api.rs
@@ -2278,10 +2278,12 @@ pub trait WalletRead {
 
     /// Returns a [`WalletSummary`] that represents the sync status and the wallet balances as of
     /// the chain tip given the specified confirmation policy for all accounts known to the wallet,
-    /// or `Ok(None)` if the wallet has no summary data available.
+    /// or `Ok(None)` if the wallet has no summary data available. Value that `capability` does
+    /// not authorize is reported as watch-only value.
     fn get_wallet_summary(
         &self,
         confirmations_policy: ConfirmationsPolicy,
+        capability: &SpendCapability<Self::AccountId>,
     ) -> Result<Option<WalletSummary<Self::AccountId>>, Self::Error>;
 
     /// Returns the height of the chain as known to the wallet as of the most recent call to
@@ -2499,13 +2501,15 @@ pub trait WalletRead {
 
     /// Returns a mapping from each transparent receiver associated with the specified account
     /// to the key scope for that address and the balance of funds given the specified target
-    /// height and confirmations policy.
+    /// height and confirmations policy. Value that `capability` does not authorize is reported
+    /// as watch-only value.
     #[cfg(feature = "transparent-inputs")]
     fn get_transparent_balances(
         &self,
         _account: Self::AccountId,
         _target_height: TargetHeight,
         _confirmations_policy: ConfirmationsPolicy,
+        _capability: &SpendCapability<Self::AccountId>,
     ) -> Result<TransparentBalances, Self::Error> {
         unimplemented!(
             "WalletRead::get_transparent_balances must be overridden for wallets to use the `transparent-inputs` feature"