+++ zcash_client_backend/src/data_api/testing.rs
@@ -1630,7 +1630,7 @@ where
     ) -> T {
         let binding = self
             .wallet()
-            .get_wallet_summary(confirmations_policy)
+            .get_wallet_summary(confirmations_policy, &self.full_spend_capability())
             .unwrap()
             .unwrap();
         f(binding.account_balances().get(&account).unwrap())
@@ -1692,7 +1692,7 @@ where
         confirmations_policy: ConfirmationsPolicy,
     ) -> Option<WalletSummary<AccountIdT>> {
         self.wallet()
-            .get_wallet_summary(confirmations_policy)
+            .get_wallet_summary(confirmations_policy, &self.full_spend_capability())
             .unwrap()
     }
 }
@@ -3385,6 +3385,7 @@ impl WalletRead for MockWalletDb {
     fn get_wallet_summary(
         &self,
         _confirmations_policy: ConfirmationsPolicy,
+        _capability: &SpendCapability<Self::AccountId>,
     ) -> Result<Option<WalletSummary<Self::AccountId>>, Self::Error> {
         Ok(None)
     }
@@ -3481,6 +3482,7 @@ impl WalletRead for MockWalletDb {
         _account: Self::AccountId,
         _target_height: TargetHeight,
         _confirmations_policy: ConfirmationsPolicy,
+        _capability: &SpendCapability<Self::AccountId>,
     ) -> Result<TransparentBalances, Self::Error> {
         Ok(HashMap::new())
     }