+++ zcash_client_backend/src/data_api/testing/pool/locking.rs
@@ -1200,7 +1200,12 @@ where
     // unaffected by lock state.
     let balances = st
         .wallet()
-        .get_transparent_balances(account_id, target_height, ConfirmationsPolicy::MIN)
+        .get_transparent_balances(
+            account_id,
+            target_height,
+            ConfirmationsPolicy::MIN,
+            &st.full_spend_capability(),
+        )
         .unwrap();
     let (_, bal) = balances
         .get(taddr)
@@ -1220,7 +1225,12 @@ where
     let expired_target = TargetHeight::from(height + 11);
     let balances = st
         .wallet()
-        .get_transparent_balances(account_id, expired_target, ConfirmationsPolicy::MIN)
+        .get_transparent_balances(
+            account_id,
+            expired_target,
+            ConfirmationsPolicy::MIN,
+            &st.full_spend_capability(),
+        )
         .unwrap();
     let (_, bal) = balances
         .get(taddr)
@@ -1249,7 +1259,12 @@ where
     );
     let balances = st
         .wallet()
-        .get_transparent_balances(account_id, expired_target, ConfirmationsPolicy::MIN)
+        .get_transparent_balances(
+            account_id,
+            expired_target,
+            ConfirmationsPolicy::MIN,
+            &st.full_spend_capability(),
+        )
         .unwrap();
     let (_, bal) = balances
         .get(taddr)